supprim des champs non necessaires dans subvention

+++ subventions-service/src/main/java/org/gov/moussaada/subventions_service/dao/SubventionDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SubventionDAO extends JpaRepository<Subvention,Integer> {
+public interface SubventionDAO extends JpaRepository<Subvention,Long> {
 }
