+++ subventions-service/src/main/java/org/gov/moussaada/subventions_service/service/inter/ISubvention.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 public interface ISubvention {
     ResponseEntity<?> Create(SubventionRequest subventionRequest);
-    ResponseEntity<?> Update(int id , SubventionRequest subventionRequest);
+    ResponseEntity<?> Update(Long id , SubventionRequest subventionRequest);
     ResponseEntity<?> GetAll();
-    ResponseEntity<?> GetById(int id);
-    ResponseEntity<?> RemoveSubvention(int id);
+    ResponseEntity<?> GetById(Long id);
+    ResponseEntity<?> RemoveSubvention(Long id);
 }
